Refactored task_list and added conditionally redirecting people_in_the_home task
