+++ arc_application/views/childminder_views/review.py
@@ -9,337 +9,19 @@ from django.shortcuts import render
 from django.utils.decorators import method_decorator
 from django.views import View
 
-from arc_application.models import AdultInHome
-from arc_application.forms.childminder_forms.form import PreviousRegistrationDetailsForm, ChildForm, ChildAddressForm
-from arc_application.forms.childminder_forms.form import AdultInYourHomeForm, ChildInYourHomeForm, OtherPersonPreviousRegistrationDetailsForm
-from arc_application.magic_link import generate_magic_link
-from arc_application.models import PreviousRegistrationDetails, OtherPersonPreviousRegistrationDetails
-from arc_application.models import ApplicantName, ApplicantPersonalDetails, Application, Arc, ArcComments, ChildcareType, UserDetails
-from arc_application.views.base import release_application
-from arc_application.notify import send_email
 from arc_application.decorators import group_required, user_assigned_application
+from arc_application.forms.childminder_forms.form import AdultInYourHomeForm, ChildInYourHomeForm, \
+    OtherPersonPreviousRegistrationDetailsForm
+from arc_application.forms.childminder_forms.form import PreviousRegistrationDetailsForm, ChildForm, ChildAddressForm
+from arc_application.magic_link import generate_magic_link
+from arc_application.models import AdultInHome
+from arc_application.models import ApplicantName, ApplicantPersonalDetails, Application, Arc, ArcComments, \
+    ChildcareType, UserDetails
+from arc_application.models import PreviousRegistrationDetails, OtherPersonPreviousRegistrationDetails
+from arc_application.notify import send_email
+from arc_application.views.base import release_application
 
 decorators = [login_required, group_required(settings.ARC_GROUP), user_assigned_application]
-
-
-@login_required
-@group_required(settings.ARC_GROUP)
-@user_assigned_application
-def task_list(request):
-    """
-    Generates the full task list for ARC users
-    :param request:  The Http request directed to the view
-    :return: The task list page with the associated context
-    """
-
-    # If the user has signed in and is a member of the ARC group
-    if request.method == 'GET':
-        application_id = request.GET['id']
-        application = Application.objects.get(application_id=application_id)
-        arc_application = Arc.objects.get(application_id=application_id)
-        personal_details_record = ApplicantPersonalDetails.objects.get(application_id=application_id)
-        name_record = ApplicantName.objects.get(personal_detail_id=personal_details_record.personal_detail_id)
-        childcare_type_record = ChildcareType.objects.get(application_id=application_id)
-
-        review_fields_to_check = set_review_fields_to_check(application, childcare_type_record)
-
-        flagged_fields_to_check = set_flagged_fields_to_check(application, childcare_type_record)
-
-        review_count = sum([1 for field in review_fields_to_check if getattr(arc_application, field) == 'COMPLETED'])
-        review_count += sum([1 for field in flagged_fields_to_check if getattr(application, field)])
-
-        # Set total number of tasks in right-hand side overview box
-        number_of_tasks = set_number_of_tasks(application, childcare_type_record)
-
-        # Load review status
-        application_status_context = {
-            'application_id': application_id,
-            'application_reference': application.application_reference,
-            'login_details_status': arc_application.login_details_review,
-            'personal_details_status': arc_application.personal_details_review,
-            'your_children_status': arc_application.your_children_review,
-            'childcare_type_status': arc_application.childcare_type_review,
-            'first_aid_training_status': arc_application.first_aid_review,
-            'criminal_record_check_status': arc_application.dbs_review,
-            'childcare_training_status': arc_application.childcare_training_review,
-            'health_status': arc_application.health_review,
-            'reference_status': arc_application.references_review,
-            'people_in_home_status': arc_application.people_in_home_review,
-            'birth_day': personal_details_record.birth_day,
-            'birth_month': personal_details_record.birth_month,
-            'birth_year': personal_details_record.birth_year,
-            'first_name': name_record.first_name,
-            'middle_names': name_record.middle_names,
-            'last_name': name_record.last_name,
-            'zero_to_five': childcare_type_record.zero_to_five,
-            'five_to_eight': childcare_type_record.five_to_eight,
-            'eight_plus': childcare_type_record.eight_plus,
-            'review_count': review_count,
-            'all_complete': all_complete(application_id, False),
-            'number_of_tasks': number_of_tasks,
-            'own_children': application.own_children,
-            'working_in_other_childminder_home': application.working_in_other_childminder_home
-        }
-
-    return render(request, 'childminder_templates/task-list.html', application_status_context)
-
-
-def set_flagged_fields_to_check(application, childcare_type_record):
-    """
-    Method to set the flagged fields to check
-    :param application: an Application object
-    :param childcare_type_record: a ChildcareType object
-    :return: integer
-    """
-    # Default flagged fields to check
-    flagged_fields_to_check = (
-        "childcare_type_arc_flagged",
-        "criminal_record_check_arc_flagged",
-        "childcare_training_arc_flagged",
-        "first_aid_training_arc_flagged",
-        "login_details_arc_flagged",
-        "personal_details_arc_flagged",
-        "references_arc_flagged"
-    )
-
-    if childcare_type_record.zero_to_five and not application.own_children and application.working_in_other_childminder_home:
-        flagged_fields_to_check = (
-            "childcare_type_arc_flagged",
-            "criminal_record_check_arc_flagged",
-            "childcare_training_arc_flagged",
-            "first_aid_training_arc_flagged",
-            "health_arc_flagged",
-            "login_details_arc_flagged",
-            "personal_details_arc_flagged",
-            "references_arc_flagged"
-        )
-    elif childcare_type_record.zero_to_five and application.own_children and application.working_in_other_childminder_home:
-        flagged_fields_to_check = (
-            "childcare_type_arc_flagged",
-            "criminal_record_check_arc_flagged",
-            "childcare_training_arc_flagged",
-            "first_aid_training_arc_flagged",
-            "health_arc_flagged",
-            "login_details_arc_flagged",
-            "personal_details_arc_flagged",
-            "references_arc_flagged",
-            'your_children_arc_flagged'
-        )
-    elif childcare_type_record.zero_to_five and application.own_children and not application.working_in_other_childminder_home:
-        flagged_fields_to_check = (
-            "childcare_type_arc_flagged",
-            "criminal_record_check_arc_flagged",
-            "childcare_training_arc_flagged",
-            "first_aid_training_arc_flagged",
-            "health_arc_flagged",
-            "login_details_arc_flagged",
-            "people_in_home_arc_flagged",
-            "personal_details_arc_flagged",
-            "references_arc_flagged",
-            'your_children_arc_flagged'
-        )
-    elif not childcare_type_record.zero_to_five and not application.own_children and application.working_in_other_childminder_home:
-        flagged_fields_to_check = (
-            "childcare_type_arc_flagged",
-            "criminal_record_check_arc_flagged",
-            "childcare_training_arc_flagged",
-            "first_aid_training_arc_flagged",
-            "login_details_arc_flagged",
-            "personal_details_arc_flagged",
-            "references_arc_flagged"
-        )
-    elif not childcare_type_record.zero_to_five and not application.own_children and not application.working_in_other_childminder_home:
-        flagged_fields_to_check = (
-            "childcare_type_arc_flagged",
-            "criminal_record_check_arc_flagged",
-            "childcare_training_arc_flagged",
-            "first_aid_training_arc_flagged",
-            "login_details_arc_flagged",
-            "people_in_home_arc_flagged",
-            "personal_details_arc_flagged",
-            "references_arc_flagged"
-        )
-    elif not childcare_type_record.zero_to_five and application.own_children and not application.working_in_other_childminder_home:
-        flagged_fields_to_check = (
-            "childcare_type_arc_flagged",
-            "criminal_record_check_arc_flagged",
-            "childcare_training_arc_flagged",
-            "first_aid_training_arc_flagged",
-            "login_details_arc_flagged",
-            "people_in_home_arc_flagged",
-            "personal_details_arc_flagged",
-            "references_arc_flagged",
-            'your_children_arc_flagged'
-        )
-    elif childcare_type_record.zero_to_five and not application.own_children and not application.working_in_other_childminder_home:
-        flagged_fields_to_check = (
-            "childcare_type_arc_flagged",
-            "criminal_record_check_arc_flagged",
-            "childcare_training_arc_flagged",
-            "first_aid_training_arc_flagged",
-            "health_arc_flagged",
-            "login_details_arc_flagged",
-            "people_in_home_arc_flagged",
-            "personal_details_arc_flagged",
-            "references_arc_flagged"
-        )
-    elif not childcare_type_record.zero_to_five and application.own_children and application.working_in_other_childminder_home:
-        flagged_fields_to_check = (
-            "childcare_type_arc_flagged",
-            "criminal_record_check_arc_flagged",
-            "childcare_training_arc_flagged",
-            "first_aid_training_arc_flagged",
-            "login_details_arc_flagged",
-            "personal_details_arc_flagged",
-            "references_arc_flagged",
-            'your_children_arc_flagged'
-        )
-
-    return flagged_fields_to_check
-
-
-def set_review_fields_to_check(application, childcare_type_record):
-    """
-    Method to set the review fields to check
-    :param application: an Application object
-    :param childcare_type_record: a ChildcareType object
-    :return: integer
-    """
-    # Default review fields to check
-    review_fields_to_check = (
-        'login_details_review',
-        'personal_details_review',
-        'childcare_type_review',
-        'first_aid_review',
-        'dbs_review',
-        'childcare_training_review',
-        'references_review'
-    )
-
-    if childcare_type_record.zero_to_five and not application.own_children and application.working_in_other_childminder_home:
-        review_fields_to_check = (
-            'login_details_review',
-            'personal_details_review',
-            'childcare_type_review',
-            'first_aid_review',
-            'dbs_review',
-            'childcare_training_review',
-            'health_review',
-            'references_review'
-        )
-    elif childcare_type_record.zero_to_five and application.own_children and application.working_in_other_childminder_home:
-        review_fields_to_check = (
-            'login_details_review',
-            'personal_details_review',
-            'childcare_type_review',
-            'first_aid_review',
-            'dbs_review',
-            'childcare_training_review',
-            'health_review',
-            'references_review',
-            'your_children_review'
-        )
-    elif childcare_type_record.zero_to_five and application.own_children and not application.working_in_other_childminder_home:
-        review_fields_to_check = (
-            'login_details_review',
-            'personal_details_review',
-            'childcare_type_review',
-            'first_aid_review',
-            'dbs_review',
-            'childcare_training_review',
-            'health_review',
-            'references_review',
-            'your_children_review',
-            'people_in_home_review'
-        )
-    elif not childcare_type_record.zero_to_five and not application.own_children and application.working_in_other_childminder_home:
-        review_fields_to_check = (
-            'login_details_review',
-            'personal_details_review',
-            'childcare_type_review',
-            'first_aid_review',
-            'dbs_review',
-            'childcare_training_review',
-            'references_review'
-        )
-    elif not childcare_type_record.zero_to_five and not application.own_children and not application.working_in_other_childminder_home:
-        review_fields_to_check = (
-            'login_details_review',
-            'personal_details_review',
-            'childcare_type_review',
-            'first_aid_review',
-            'dbs_review',
-            'childcare_training_review',
-            'references_review',
-            'people_in_home_review'
-        )
-    elif not childcare_type_record.zero_to_five and application.own_children and not application.working_in_other_childminder_home:
-        review_fields_to_check = (
-            'login_details_review',
-            'personal_details_review',
-            'childcare_type_review',
-            'first_aid_review',
-            'dbs_review',
-            'childcare_training_review',
-            'references_review',
-            'your_children_review',
-            'people_in_home_review'
-        )
-    elif childcare_type_record.zero_to_five and not application.own_children and not application.working_in_other_childminder_home:
-        review_fields_to_check = (
-            'login_details_review',
-            'personal_details_review',
-            'childcare_type_review',
-            'first_aid_review',
-            'dbs_review',
-            'childcare_training_review',
-            'health_review',
-            'references_review',
-            'people_in_home_review'
-        )
-    elif not childcare_type_record.zero_to_five and application.own_children and application.working_in_other_childminder_home:
-        review_fields_to_check = (
-            'login_details_review',
-            'personal_details_review',
-            'childcare_type_review',
-            'first_aid_review',
-            'dbs_review',
-            'childcare_training_review',
-            'references_review',
-            'your_children_review'
-        )
-
-    return review_fields_to_check
-
-
-def set_number_of_tasks(application, childcare_type_record):
-    """
-    Method to set the total number of tasks in the right hand summary panel
-    :param application: an Application object
-    :param childcare_type_record: a ChildcareType object
-    :return: integer
-    """
-    # Default total number of tasks
-    number_of_tasks = 7
-
-    if childcare_type_record.zero_to_five and not application.own_children and application.working_in_other_childminder_home:
-        number_of_tasks = 8
-    elif childcare_type_record.zero_to_five and application.own_children and application.working_in_other_childminder_home:
-        number_of_tasks = 9
-    elif childcare_type_record.zero_to_five and application.own_children and not application.working_in_other_childminder_home:
-        number_of_tasks = 10
-    elif not childcare_type_record.zero_to_five and not application.own_children and application.working_in_other_childminder_home:
-        number_of_tasks = 7
-    elif not childcare_type_record.zero_to_five and not application.own_children and not application.working_in_other_childminder_home:
-        number_of_tasks = 8
-    elif not childcare_type_record.zero_to_five and application.own_children and not application.working_in_other_childminder_home:
-        number_of_tasks = 9
-    elif childcare_type_record.zero_to_five and not application.own_children and not application.working_in_other_childminder_home:
-        number_of_tasks = 9
-    elif not childcare_type_record.zero_to_five and application.own_children and application.working_in_other_childminder_home:
-        number_of_tasks = 8
-
-    return number_of_tasks
 
 
 @group_required(settings.ARC_GROUP)
@@ -424,32 +106,6 @@ def has_group(user, group_name):
     group = Group.objects.get(name=group_name)
 
     return True if group in user.groups.all() else False
-
-
-def all_complete(id, flag):
-    """
-    Check the status of all sections
-    :param id: Application Id
-    :return: True or False depending on whether all sections have been reviewed
-    """
-    if Arc.objects.filter(application_id=id):
-        arc = Arc.objects.get(application_id=id)
-        tasks = [arc.login_details_review, arc.childcare_type_review, arc.personal_details_review,
-                 arc.first_aid_review, arc.childcare_training_review, arc.dbs_review, arc.people_in_home_review]
-        zero_to_five = ChildcareType.objects.get(application_id=id).zero_to_five
-        if zero_to_five:
-            tasks.append(arc.health_review)
-            tasks.append(arc.references_review)
-        application = Application.objects.get(application_id=id)
-        if application.own_children is True:
-            tasks.append(arc.your_children_review)
-
-        for i in tasks:
-
-            if (i == 'NOT_STARTED' and not flag) or (i != 'COMPLETED' and flag):
-                return False
-
-        return True
 
 
 # Add personalisation and create template
